requirements.txt
langchain
langchain-core
langchain-community
langchain-ollama
python-dotenv
requests
langchain-core
python-dotenv
groq
langchain-groq
fastapi
uvicorn
langchain_aws
